requirements.txt
streamlit
google-cloud-aiplatform
google-cloud-texttospeech
langgraph
langchain
langchain-google-vertexai
pydantic
